Internalize queries that are being used in jobs
The *Snapshot queries are being used in the display-activity-extractor
and have been internalized in that repo so that changes to the queries
can't be changed outside of version control and the change requires a
review.

The brandPromiseKPIs, userChanges, and companyChanges queries are used in the
hubspot-connector and internalized there.

The subscriptionChanges query is left in this repo since it is not used
in any job at this point.